average 동작 수정
