+++ test/__tests__/Index.mjs
@@ -1,20 +1,21 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
+import * as Math_test from "./spec/Math_test.mjs";
 import * as Array_test from "./spec/Array_test.mjs";
 import * as Object_test from "./spec/Object_test.mjs";
 import * as String_test from "./spec/String_test.mjs";
 import * as Array_NonEmpty_test from "./spec/Array_NonEmpty_test.mjs";
 
-var testEqual = Array_NonEmpty_test.testEqual;
-
 var emptyArray = Array_NonEmpty_test.emptyArray;
 
 var nonEmptyArray = Array_NonEmpty_test.nonEmptyArray;
 
+var testEqual = String_test.testEqual;
+
 export {
-  testEqual ,
   emptyArray ,
   nonEmptyArray ,
+  testEqual ,
   
 }
-/* Array_test Not a pure module */
+/* Math_test Not a pure module */
